Comments out broken test until we fix it
I gave this a shot for a few hours on a slow shift. I added sinon and another library to mock request animation frame, but I didn't have any luck. I wasn't able to get the spec to behave the way I wanted it to (it was either always failing or always passing). For now, it seems like we should comment out this test until we fix it. Otherwise every student who gets to this point is going to stop and ask a question.